+++ tests/test_utils.py
@@ -221,13 +221,6 @@ def test_get_all_git_refs(tmpdir):
     assert utils.get_all_git_refs(repo_dir=tmpdir) == ['master']
 
 
-@mark.slow
-def test_get_sha_for_current_git_ref(tmpdir):
-    hex_sha = utils.get_sha_for_current_git_ref(git_ref='master', repo='emcek/common_sense', repo_dir=tmpdir)
-    assert len(hex_sha) == 40
-    assert hex_sha == '6d5d9f80a8309eec3e9cf183fc57d56559035560'
-
-
 def test_check_dcs_bios_entry_no_entry(tmpdir):
     from os import makedirs
     install_dir = tmpdir / 'install'
@@ -313,7 +306,7 @@ def test_run_pip_command_failed():
 @mark.skipif(condition=platform != 'win32', reason='Run only on Windows')
 @mark.parametrize('cmd, result', [('Clear-Host', 0), ('bullshit', -1)])
 def test_run_command(cmd, result):
-    rc = utils.run_command(cmd=f'powershell {cmd}')
+    rc = utils.run_command(cmd=['powershell', cmd])
     assert rc == result